% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compact2long}
\alias{compact2long}
\title{compact2long}
\usage{
compact2long(infile)
}
\arguments{
\item{infile}{file path to input file (xls[x])}
}
\description{
Converts tables in compact format to long format
}
\details{
The following standard names in the data field of the compact form are reserved
\itemize{
  \item{Year, Month, Day, Time:}{date indicators}
  \item{P:}{pressure reading (in mm, in, or other)}
  \item{QFE:}{station pressure (in hPa)}
  \item{QFF:}{pressure at sea level}
  \item{TA:}{outside temperature}
  \item{TP:}{temperature at barometer}
}
}
\keyword{util}

