% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_time}
\alias{convert_time}
\title{convert_time}
\usage{
convert_time(year, month, day, time, latitude = NULL)
}
\arguments{
\item{year}{numeric or character, year corresponding to time}

\item{month}{numeric or character, month of time}

\item{day}{numeric or character, day of time}

\item{time}{character vector of representations of time}
}
\description{
Converts time in various formats to HH:MM
}
\details{
Conversion of time to times in HH:MM format is based on the solar declination
from the insol package if times specify 'sunset' or 'sunrise'. In addition,
the string representations of 'morning', 'noon', 'afternoon', 'evening' are
set (see example(convert_time))
}
\examples{
time <- c('12PM', '12AM', 'sunrise', 'sunset','morning', 'noon',  'afternoon', 'evening', '12:00', '1:00', '13:13', '04:00')
cbind(time, convert_time(1815, 1, 1, time, latitude=45))
}
\keyword{util}

