% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expand_long}
\alias{expand_long}
\title{expand_long}
\usage{
expand_long(df, inventory = read_inventory(), verbose = TRUE)
}
\arguments{
\item{df}{input data frame in long format (standard names)}

\item{inventory}{data frame containing station information available in inventory}

\item{verbose}{logical, should information about processing steps be displayed?}
}
\description{
Expands the long format to fill missing columns
}
\keyword{util}

