% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_T20CR}
\alias{get_T20CR}
\title{get_T20CR}
\usage{
get_T20CR(lon, lat, time, tfile = "~/Unibe/20CR/tmean.2m.1871-1900.nc",
  tfilter = rep(1/11, 11))
}
\arguments{
\item{lon}{longitude of station in degrees}

\item{lat}{latitude of station in degrees}

\item{time}{of measurement in UTC}

\item{tfile}{file path to temperature climatologies}

\item{tfilter}{filter weights to smooth temperature climatology}
}
\description{
get Temperature climatology from 20CR
}
\keyword{util}

