% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gravity}
\alias{gravity}
\title{gravity}
\usage{
gravity(phi, elevation = 0)
}
\arguments{
\item{phi}{latitude (in degrees)}

\item{elevation}{elevation (in m)}
}
\description{
Compute gravity at latitude phi
}
\examples{
## gravity at equator
gravity(0)
## gravity at pole
gravity(90)
}
\references{
http://library.wmo.int/pmb_ged/wmo_8_en-2012.pdf
}
\keyword{util}

