% Generated by roxygen2 (4.0.1): do not edit by hand
\name{length2SI}
\alias{length2SI}
\title{length2SI}
\usage{
length2SI(x, units = "imperial", base = NULL, verbose = TRUE)
}
\arguments{
\item{x}{vector, or matrix of lengths}

\item{units}{length units to be used (see details)}

\item{base}{conversion factors for columns of matrix if non-standard
units are used, character if in base list}

\item{verbose}{logical, should information on conversion factors be displayed?}
}
\description{
Convert length units from non-standard units to SI.
This allows to convert imperial units to m, but also using
various bases (some of which are predefined, some of which
have to be supplied by the user).
}
\examples{
length2SI(rbind(c(20,3,2), c(0,6,7)))
length2SI(1, 'finger')
}
\references{
http://www.thefullwiki.org/Conversion_of_units
}
\keyword{util}

