% Generated by roxygen2 (4.0.1): do not edit by hand
\name{local2UTC}
\alias{local2UTC}
\title{local2UTC}
\usage{
local2UTC(x, longitude = 0)
}
\arguments{
\item{x}{a datum (local time) of class POSIXct to be converted}

\item{longitude}{longitude of the location in degree at which local time is
supplied. If no longitude is supplied the same datum should be returned.}
}
\description{
Converts local (solar) time to UTC based on simple longitude correction.
So far, this conversion does not take into account the varying length of
solar days throughout the year.
}
\examples{
dd <- as.POSIXct('1815-01-01 12:00:01', tz='UTC')
local2UTC(dd)

# not setting the time zone to UTC throws a warning
local2UTC(as.POSIXct('1815-01-01 12:00:01'))

## noon at the date line is midnight in England
local2UTC(dd, 180)
local2UTC(dd, -180)
}
\keyword{util}

