% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pressure2SI}
\alias{pressure2SI}
\title{pressure2SI}
\usage{
pressure2SI(pressure, temperature, latitude, elevation = 0)
}
\arguments{
\item{pressure}{Local absolute pressure in mmHg}

\item{temperature}{Local temperature in deg. C}

\item{latitude}{latitude of location in degrees for gravity}

\item{elevation}{station elevation in m for local gravity (defaults to 0)}
}
\description{
Converts the absolute pressure in mmHg to hPa given the local temperature
}
\examples{
## should be 1000 hPa
pressure2SI(750, temperature=10, latitude=68.61)
}
\keyword{util}

