% Generated by roxygen2 (4.0.1): do not edit by hand
\name{QFE2QFF}
\alias{QFE2QFF}
\title{QFE2QFF}
\usage{
QFE2QFF(QFE, elevation, latitude, temperature)
}
\arguments{
\item{QFE}{local atmospheric pressure (in hPa)}

\item{elevation}{height AMSL of barometer (in m)}

\item{latitude}{latitude of station (in degree)}

\item{temperature}{air temperature at station (in Kelvin)}
}
\description{
Reduce absolute atmospheric pressure to sea level
}
\details{
The mean temperature of the fictituous air column between station
elevation and sea level is estimated using a lapse rate of 0.0065 K/m.
}
\examples{
## should be a 1000 hPa at sea level
QFE2QFF(990, elevation=88, latitude=15, temperature=298.15)
QFE2QFF(990, elevation=88, latitude=15, temperature=25)
}
\keyword{util}

