% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_inventory}
\alias{read_inventory}
\title{read_inventory}
\usage{

  read_inventory(file = "~/Unibe/pressure/Working_inventory_pressure_1815-17.xls")
}
\arguments{
\item{file}{path to file containing inventory information}
}
\description{
Reads metadata in common repository for later use
}
\keyword{util}

