% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_pressure}
\alias{read_pressure}
\title{read_pressure}
\usage{
read_pressure(station, datapath = "~/Unibe/pressure/orig_data")
}
\arguments{
\item{station}{name of the station}

\item{datapath}{file path with station files}
}
\description{
Read in the original (digitized) pressure series at different stations
}
\details{
This is a simple wrapper function to read in the station data from Excel/ASCII.
For each new format (almost for every station), a new function exists. As this
may be unwieldy for short files, alternatively the original data may be manually
copied to the standard format.

The idea is to reorganise the original data to a long format (one observing time
per line) and convert the units to SI units in functions specific to each station.
From there, adding longitudes and latitudes, temperature (from climatology if not
in original files) and correcting to standard temperature (0 deg. C) and sea level
will be done automatically for all data read in.
}
\keyword{util}

