% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_Sweden}
\alias{read_Sweden}
\title{read_Sweden}
\usage{
read_Sweden(station, infile)
}
\arguments{
\item{station}{Station name}

\item{infile}{file path of air pressure data}
}
\description{
Reads in Swedish air pressure station data and transforms units
}
\keyword{util}

