% Generated by roxygen2 (4.0.1): do not edit by hand
\name{temperature2SI}
\alias{temperature2SI}
\title{temperature2SI}
\usage{
temperature2SI(x, units = "C")
}
\arguments{
\item{x}{temperature values}

\item{units}{temperature units}
}
\description{
Convert temperature from non-standard units to SI. By default,
this function converts degrees Celsius to Kelvin.
}
\examples{
x <- c(0, 32, 100)
temperature2SI(x, 'F')

temperature2SI(26.85)
}
\references{
http://www.thefullwiki.org/Conversion_of_units
}
\keyword{util}

