% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gravity}
\alias{gravity}
\title{gravity}
\usage{
gravity(phi)
}
\arguments{
\item{phi}{latitude (in degrees)}
}
\description{
Compute gravity at latitude phi
}
\examples{
## gravity at equator
gravity(0)
## gravity at pole
gravity(90)
}
\keyword{util}

